requirements.txt
numpy
opencv-python
scikit-learn
matplotlib
pillow
torch
torchvision
tqdm